Fix test compatibility
First, unittest2 is not compatible with python 3.10. However, in Python
3.x, the standard unittest module has the functionality we used, so
unittest2 is now only used in python 2.7.

Next. changed assertItemsEqual to AssertEqual. It seems that
assertItemsEqual only compared item counts, not the instances, and was
removed due to the majority of users, this project included, assuming
that the instances were compared. See https://bugs.python.org/issue27060